Пагинация, сортировка, фильтрация

+++ Keeper.Domen/Models/User.cs
@@ -8,11 +8,10 @@ namespace Keeper.Domen.Models;
 
 public class User
 {
-    public long Id { get; set; }
+    public int Id { get; set; }
     public string Email { get; set; }
     public string Login { get; set; }
     public string Password { get; set; }
-
     public IEnumerable<Statement>? Statements { get; set; } = null;
 
 }
